reoranized endpoints to auth
